Reverse changes not intended to be part of distillation release
